- Issues 59-61 And Customer/Deliverer token

Closes #48

See merge request ase-22-23/team40/ase-project-2022-team-40!28

+++ microservices/delivery-service/src/main/java/com/group40/deliveryservice/controller/UserController.java
@@ -1,17 +1,20 @@
 package com.group40.deliveryservice.controller;
 
+import com.group40.deliveryservice.dto.PersonResponse;
 import com.group40.deliveryservice.dto.UserRequest;
-import com.group40.deliveryservice.model.Customer;
-import com.group40.deliveryservice.model.Deliverer;
-import com.group40.deliveryservice.model.Dispatcher;
-import com.group40.deliveryservice.model.User;
+import com.group40.deliveryservice.model.*;
 import com.group40.deliveryservice.repository.UserRepository;
 import com.group40.deliveryservice.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.json.JSONException;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.IOException;
+import java.security.SecureRandom;
+import java.util.Base64;
 import java.util.List;
 
 
@@ -24,25 +27,43 @@ public class UserController {
 
     private final UserRepository userRepository;
 
+    private final SecureRandom secureRandom;
+
+    private final Base64.Encoder base64Encoder;
+
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    List<Customer> getAllCustomers() {
-        return userService.getAllCustomers();
+    ResponseEntity<?> getAllCustomers(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            return ResponseEntity.ok(userService.getAllCustomers());
+        }else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @DeleteMapping("/{id}")
-    void deleteUser(@PathVariable String id) {
-        userService.deleteUser(id);
+    ResponseEntity<?> deleteUser(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable String id) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            userService.deleteUser(id);
+            return ResponseEntity.ok().build();
+        }else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<?> updateUser(@RequestBody UserRequest newUser, @PathVariable String id) {
+    ResponseEntity<?> updateUser(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody UserRequest newUser, @PathVariable String id) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (!user.getRole().equals(ERole.ROLE_DISPATCHER) && !user.getId().equals(id)) {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
         switch (newUser.getRole()) {
             case ROLE_DELIVERER -> {
                 Deliverer deliverer = Deliverer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
-                        .deliveries(newUser.getDeliveries())
                         .build();
                 return ResponseEntity.ok(userService.updateUser(deliverer, id));
             }
@@ -57,8 +78,6 @@ public class UserController {
                 Customer customer = Customer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
-                        .boxes(newUser.getBoxes())
-                        .deliveries(newUser.getDeliveries())
                         .build();
                 return ResponseEntity.ok(userService.updateUser(customer, id));
             }
@@ -67,18 +86,20 @@ public class UserController {
 
     @PostMapping
     ResponseEntity<?> createUser(@RequestBody UserRequest newUser) {
-        //TODO: Better handling
         if (userRepository.existsByEmail(newUser.getEmail())) {
             throw new RuntimeException("Email already in use!");
         }
         if(newUser.getPassword() == null || newUser.getPassword().isBlank() || newUser.getPassword().trim().isBlank()){
             throw new RuntimeException("Creating a new user requires a password.");
         }
+        byte[] randomBytes = new byte[64];
+        secureRandom.nextBytes(randomBytes);
         switch (newUser.getRole()) {
             case ROLE_DELIVERER -> {
                 Deliverer deliverer = Deliverer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
+                        .token(base64Encoder.encodeToString(randomBytes))
                         .build();
                 return ResponseEntity.ok(userService.createUser(deliverer, newUser.getPassword()));
             }
@@ -93,8 +114,26 @@ public class UserController {
                 Customer customer = Customer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
+                        .token(base64Encoder.encodeToString(randomBytes))
                         .build();
                 return ResponseEntity.ok(userService.createUser(customer, newUser.getPassword()));
+            }
+        }
+    }
+
+    @GetMapping("/token")
+    public ResponseEntity<?> getUserToken(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        switch (user.getRole()){
+            case ROLE_DELIVERER -> {
+                return ResponseEntity.ok(((Deliverer) user).getToken());
+            }
+            case ROLE_CUSTOMER -> {
+                return ResponseEntity.ok(((Customer) user).getToken());
+            }
+            default -> {
+                return ResponseEntity.badRequest()
+                        .body("Incorrect role");
             }
         }
     }
